+++ CHANGES.md
@@ -7,6 +7,7 @@ CHANGELOG
 **Bug fixes**
 
 - Fix Log Entry GeoJSON with no geometry
+- Fix log entry access by creating mapentity.read_logentry permission
 - Hide log entry menu for users without permission.
 - Hide "Show full history" from object for users without permission.
 - Hide admin menu entry for users without permission.